Fixed authentication
