+++ src/services/helper/Parser.ts
@@ -1,31 +1,3 @@
-// This file contains various methods for extracting data from incoming JSON
-
-// TYPES
-import { AxiosResponseHeaders } from 'axios';
-
-/**
- * @returns The cookie string after removing unnecessary data from it
- * @param headers The raw header from which the cookies are to be extracted
- */
-export function parseCookies(headers: AxiosResponseHeaders): string {
-    // Getting the raw cookie string
-    var cookies: string = headers['set-cookie'] + '';
-    var newCookie: string = '';
-
-    // Getting required cookies from all cookies
-    newCookie += cookies.match(/guest_id_marketing.+?;/) + " ";
-    newCookie += cookies.match(/guest_id_ads.+?;/) + " ";
-    newCookie += cookies.match(/kdt.+?;/) + " ";
-    newCookie += cookies.match(/personalization_id.+?;/) + " ";
-    newCookie += cookies.match(/guest_id.+?;/) + " ";
-    newCookie += cookies.match(/twid.+?;/) + " ";
-    newCookie += cookies.match(/guest_id_marketing.+?;/) + " ";
-    newCookie += cookies.match(/ct0.+?;/) + " ";
-    newCookie += cookies.match(/auth_token.+?;/) + " ";
-
-    return newCookie;
-}
-
 /**
  * @returns Whether the given json object is empty or not
  * @param data The input JSON object which needs to be checked
