+++ src/types/HTTP.ts
@@ -9,15 +9,6 @@ export enum HttpMethods {
 };
 
 /**
- * @summary Stores the different types of authentication available to authencticate HTTP requests
- */
-export enum AuthType {
-    "NONE",
-    "GUEST",
-    "AUTH"
-};
-
-/**
  * @summary Stores the different types of http status codes
  */
 export enum HttpStatus {
