--- src/types/Logs.ts
@@ -1,8 +1,0 @@
-/**
- * @summary Stores a single log entry
- */
-export interface Log {
-    time: Date
-    message: string
-    data: JSON
-}
